project/requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
amqp==2.6.0
APScheduler==3.6.3
asgiref==3.5.2
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
black==22.3.0
celery==4.4.6
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
clarifai==2.6.2
click==8.1.2
colorama==0.4.3
configparser==5.0.2
constantly==15.1.0
cowsay==3.0
cryptography==2.9.2
cycler==0.10.0
daphne==2.5.0
decorator==4.4.2
Django==4.0.6
django-audiofield==0.10.0
django-braces==1.14.0
django-celery==3.3.1
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-filter==22.1
django-model-utils==4.1.1
django-money==1.1
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-private-chat==0.3.0
django-rest-framework==0.1.0
django-webpush==0.3.3
djangorestframework==3.11.0
djangorestframework-simplejwt==5.2.0
dnspython==2.0.0
EasyProcess==0.3
emoji==1.2.0
eventlet==0.25.2
future==0.18.2
googleapis-common-protos==1.53.0
greenlet==0.4.16
grpcio==1.37.0
hiredis==1.1.0
http-ece==1.1.0
hyperlink==19.0.0
idna==2.10
imageio==2.9.0
incremental==17.5.0
instapy==0.6.13
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.18.0
joblib==0.14.1
jsonschema==2.6.0
jupyter-client==6.2.0
jupyter-core==4.7.1
kiwisolver==1.1.0
kombu==4.6.11
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
MeaningCloud-python==2.0.0
monotonic==1.5
MouseInfo==0.1.3
mpmath==1.2.1
msgpack==1.0.0
multidict==4.7.6
mypy-extensions==0.4.3
nest-asyncio==1.5.1
networkx==2.5
numpy==1.18.0
oauthlib==3.1.1
opencv-python==4.4.0.46
packaging==20.9
pandas==0.25.3
parso==0.8.2
pathspec==0.9.0
phonenumbers==8.12.6
pickleshare==0.7.5
Pillow==7.1.2
pipenv==2020.6.2
platformdirs==2.5.1
plyer==2.0.0
prompt-toolkit==3.0.18
protobuf==3.15.8
psycopg2==2.8.4
py-moneyed==0.8.0
py-vapid==1.8.2
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
pyfcm==1.5.1
pygame==1.9.6
PyGetWindow==0.0.9
Pygments==2.8.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==3.13.1
python-socketio==4.6.0
python-telegram-bot==13.4.1
PyTweening==1.0.3
pytz==2020.1
PyWavelets==1.1.1
pywebpush==1.9.4
pywhatkit==4.7
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
regex==2021.4.4
requests==2.24.0
requests-oauthlib==1.3.0
scikit-image==0.18.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
service-identity==18.1.0
Shapely==1.7.1
SimpleCV==1.3
six==1.13.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.3.1
sympy==1.8
tifffile==2021.1.8
tomli==2.0.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
tweepy==3.10.0
twilio==6.53.0
Twisted==20.3.0
txaio==20.4.1
typing-extensions==4.1.1
tzdata==2022.1
tzlocal==2.1
urllib3==1.25.10
validate-email==1.3
vine==1.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.5
wcwidth==0.2.5
webdriverdownloader==1.1.0.3
websocket-client==0.57.0
websockets==8.1
wikipedia==1.4.0
wrapt==1.11.2
xlrd==1.2.0
yarl==1.4.2
zope.interface==5.1.0
